requirements.txt
APScheduler==3.10.4
beautifulsoup4==4.12.3
blinker==1.8.2
cachelib==0.9.0
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
dash==2.17.0
dash-bootstrap-components==1.6.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dateparser==1.2.0
exchange_calendars==4.5.4
Flask==3.0.3
Flask-Caching==2.3.0
gunicorn==22.0.0
idna==3.7
importlib_metadata==7.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
korean-lunar-calendar==0.3.1
llvmlite==0.42.0
lxml==4.9.4
MarkupSafe==2.1.5
nest-asyncio==1.6.0
numba==0.59.1
numpy==1.26.4
orjson==3.10.3
packaging==24.1
pandas==2.1.4
plotly==5.22.0
polygon-api-client==1.13.7
pyluach==2.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
regex==2024.5.15
requests==2.32.3
requests-futures==1.0.1
retrying==1.3.4
scipy==1.13.0
six==1.16.0
soupsieve==2.5
tenacity==8.3.0
toolz==0.12.1
tqdm==4.66.4
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
urllib3==1.26.19
websockets==12.0
Werkzeug==3.0.3
yahooquery==2.3.7
zipp==3.19.2
